refactoring components
Added components folder
Added App, LableComp, and Timer files for components
